Refactor conversion clients classes - move common code to base_conversion_client
